Optimal Approach
Best Way To Solve The Second And Largest Element in the Sorted,unique and non - neagtive ELements....
 if (arr1[i] > largest_Element) {
            Second_Largest = largest_Element;
            largest_Element = arr1[i];
            
        } else if (arr1[i]<largest_Element && arr1[i] > Second_Largest) {
            Second_Largest = arr1[i];
        }
This else if condition is inside a loop iterating over the elements of the array arr1. It's designed to update the Second_Largest variable  if the current element (arr1[i]) is less than the largest_Element but greater than the current Second_Largest element found so far.